# openapi/v2.yaml
openapi: 3.1.0
info:
  title: Sellchaze API
  version: 2.0.0-alpha.1
  description: Organization-scoped API. API v1 remains compatible during migration.
servers:
  - url: /api/v2
security:
  - bearerAuth: []
paths:
...
  /organizations/{organization}/stores:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
    get:
      operationId: listOrganizationStores
      responses:
        '200': { description: Paginated stores }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createOrganizationStore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name: { type: string, maxLength: 160 }
                slug: { type: string, maxLength: 160, pattern: '^[A-Za-z0-9_-]+$' }
                description: { type: [string, 'null'] }
                email: { type: [string, 'null'], format: email }
                phone: { type: [string, 'null'] }
                currency: { type: string, maxLength: 8 }
      responses:
        '201': { description: Store created }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /organizations/{organization}/stores/{store}/themes:
    parameters: &organizationStoreParameters
      - $ref: '#/components/parameters/OrganizationId'
      - $ref: '#/components/parameters/StoreId'
    get:
      operationId: listOrganizationStoreThemes
      responses: { '200': { description: Installed and available themes } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OrganizationId:
      name: organization
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
...
    StoreId:
      name: store
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
...
  responses:
...
    Forbidden:
      description: Membership or role does not permit this operation
    ValidationError:
      description: Validation failed